criação do banco de dados e logicas de atualizar, e setar as vagas como ocupadas ou vazias

+++ src/main/java/br/com/estacionamento/controllers/response/ComprovanteResponse.java
@@ -7,15 +7,15 @@ public class ComprovanteResponse {
     private String modelo;
     private String placa;
     private LocalDateTime dataEHora;
-    private int numVaga;
+    private long numVaga;
 
     public ComprovanteResponse() {}
 
-    public int getNumVaga() {
+    public long getNumVaga() {
         return numVaga;
     }
 
-    public void setNumVaga(int numVaga) {
+    public void setNumVaga(long numVaga) {
         this.numVaga = numVaga;
     }
