+++ src/main/java/br/com/estacionamento/entities/Comprovante.java
@@ -7,17 +7,17 @@ public class Comprovante {
     private String modelo;
     private String placa;
     private LocalDateTime dataEHora;
-    private int numVaga;
+    private long numVaga;
 
 
     public Comprovante() {}
 
 
-    public int getNumVaga() {
+    public long getNumVaga() {
         return numVaga;
     }
 
-    public void setNumVaga(int numVaga) {
+    public void setNumVaga(long numVaga) {
         this.numVaga = numVaga;
     }
